fetch/requirements.txt
python_dateutil==2.9.0.post0
Requests==2.32.4
dotenv
